Move code from HttpExceptionHandler to `ExceptionHandler`
